feat: 【RUM 检索】字段统计分析弹层组件目录化重构与数值/时长模式适配 --story=137987421
