--- bkmonitor/webpack/src/trace/pages/trace-explore/components/statistics-list.scss
@@ -1,329 +1,0 @@
-.trace-explore-dimension-statistics-popover {
-  .skeleton-wrap {
-    .skeleton-element {
-      height: 28px;
-      margin-top: 12px;
-    }
-  }
-
-  .info-skeleton {
-    .total-skeleton {
-      display: flex;
-      justify-content: space-between;
-      margin-bottom: 12px;
-
-      .skeleton-element {
-        width: 70px;
-        height: 20px;
-      }
-    }
-
-    .info-skeleton {
-      display: flex;
-      flex-wrap: wrap;
-      align-content: space-between;
-      justify-content: space-between;
-      height: 76px;
-      margin-bottom: 15px;
-
-      .skeleton-element {
-        width: 182px;
-        height: 36px;
-      }
-    }
-
-    .skeleton-element.chart {
-      height: 140px;
-      margin-bottom: 12px;
-    }
-  }
-
-  .statistics-info {
-    .top-k-info-header {
-      display: flex;
-      align-items: center;
-      justify-content: space-between;
-      margin-bottom: 12px;
-
-      .label-item {
-        font-size: 12px;
-        line-height: 20px;
-        color: #4d4f56;
-      }
-    }
-
-    .integer-statics-info {
-      display: flex;
-      flex-wrap: wrap;
-      align-content: space-between;
-      justify-content: space-between;
-      height: 76px;
-      margin-bottom: 15px;
-
-      .integer-item {
-        width: 182px;
-        height: 36px;
-        line-height: 36px;
-        text-align: center;
-        background: #f5f7fa;
-        border-radius: 2px;
-
-        .label {
-          font-size: 12px;
-          color: #4d4f56;
-        }
-
-        .value {
-          margin-left: 8px;
-          font-size: 12px;
-          font-weight: 700;
-          color: #4d4f56;
-        }
-      }
-    }
-
-    .top-k-chart-title {
-      display: flex;
-      justify-content: space-between;
-      gap: 20px;
-      margin-bottom: 4px;
-      line-height: 22px;
-
-      .title {
-        flex-shrink: 0;
-        font-size: 14px;
-        color: #313238;
-      }
-
-      .time-range {
-        overflow: hidden;
-        text-overflow: ellipsis;
-        white-space: nowrap;
-        font-size: 12px;
-        color: #4d4f56;
-      }
-    }
-  }
-
-  .top-k-list-header {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    height: 22px;
-
-    .field-name {
-      font-size: 12px;
-    }
-
-    .desc {
-      font-size: 12px;
-    }
-
-    .loading-icon {
-      width: 20px;
-      height: 20px;
-    }
-
-    .download-tool {
-      margin-left: 10px;
-      cursor: pointer;
-
-      .icon-xiazai2 {
-        font-size: 14px;
-        color: #4d4f56;
-
-        &:hover {
-          color: #3a84ff;
-        }
-      }
-    }
-  }
-
-  .load-more {
-    margin: 8px 0 0 46px;
-    font-size: 12px;
-    line-height: 20px;
-    color: #3a84ff;
-    cursor: pointer;
-
-    &.is-duration {
-      margin-left: 20px;
-    }
-  }
-
-  .empty-text {
-    font-size: 12px;
-  }
-}
-
-.trace-dimension-top-k-slider {
-  .bk-modal-mask {
-    background-color: transparent;
-  }
-
-  .bk-sideslider-title {
-    overflow: hidden;
-  }
-
-  .dimension-slider-header {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    width: 100%;
-    padding-right: 24px;
-
-    .field-name {
-      font-size: 16px;
-    }
-
-    .desc {
-      font-size: 12px;
-    }
-
-    .loading-icon {
-      width: 20px;
-      height: 20px;
-    }
-
-    .download-tool {
-      margin-left: 10px;
-      cursor: pointer;
-
-      .icon-xiazai2 {
-        font-size: 14px;
-        color: #4d4f56;
-      }
-
-      .text {
-        margin-left: 6px;
-        font-size: 12px;
-        color: #4d4f56;
-      }
-    }
-  }
-
-  .dimension-slider-content {
-    padding: 12px 40px;
-
-    .slider-load-more {
-      width: 120px;
-      height: 27px;
-      margin: 16px auto 0;
-      font-size: 12px;
-      line-height: 27px;
-      color: #3a84ff;
-      text-align: center;
-      cursor: pointer;
-      background: #f0f5ff;
-      border-radius: 14px;
-
-      &.is-loading {
-        background: transparent;
-      }
-
-      &:hover {
-        background: #e1ecff;
-
-        &.is-loading {
-          background: transparent;
-        }
-      }
-    }
-  }
-}
-
-.dimension-top-k-title {
-  display: flex;
-  flex: 1;
-  align-items: center;
-  overflow: hidden;
-
-  .field-name {
-    overflow: hidden;
-    text-overflow: ellipsis;
-    color: #313238;
-    white-space: nowrap;
-  }
-
-  .divider {
-    display: inline-block;
-    width: 1px;
-    height: 14px;
-    margin: 0 8px 0 6px;
-    background: #dcdee5;
-  }
-
-  .desc {
-    flex-shrink: 0;
-    color: #979ba5;
-  }
-}
-
-.top-k-list {
-  .top-k-list-item {
-    display: flex;
-    margin-top: 12px;
-
-    .filter-tools {
-      position: relative;
-      margin: 2px 6px 0 0;
-      color: #3a84ff;
-
-      .icon-monitor {
-        cursor: pointer;
-      }
-
-      .icon-sousuo- {
-        margin-left: 8px;
-      }
-    }
-
-    .progress-content {
-      flex: 1;
-      overflow: hidden;
-
-      .info-text {
-        display: flex;
-        align-items: center;
-        justify-content: space-between;
-        height: 20px;
-        margin-bottom: 2px;
-
-        .field-name {
-          flex: 1;
-          overflow: hidden;
-          text-overflow: ellipsis;
-          font-size: 12px;
-          color: #313238;
-          white-space: nowrap;
-
-          .sub-name {
-            color: #757880;
-          }
-        }
-
-        .counts {
-          font-size: 12px;
-          color: #5ab8a8;
-
-          .progress-count {
-            margin-left: 4px;
-            font-weight: 700;
-          }
-        }
-      }
-    }
-  }
-}
-
-.slide-dimension-filter-overflow-tips {
-  font-size: 12px;
-}
-
-.slide-dimension-filter-overflow-tips,
-.statistics-top-k-item-tooltips-wrap-popover {
-  max-width: 400px;
-  word-break: break-word;
-  overflow-wrap: break-word;
-  white-space: normal;
-}
