+++ bkmonitor/webpack/src/trace/pages/trace-explore/components/statistics-list/statistics-info.scss
@@ -1,0 +1,100 @@
+.trace-explore-dimension-statistics-popover {
+  .info-skeleton {
+    .total-skeleton {
+      display: flex;
+      justify-content: space-between;
+      margin-bottom: 12px;
+
+      .skeleton-element {
+        width: 70px;
+        height: 20px;
+      }
+    }
+
+    .info-skeleton {
+      display: flex;
+      flex-wrap: wrap;
+      align-content: space-between;
+      justify-content: space-between;
+      height: 76px;
+      margin-bottom: 15px;
+
+      .skeleton-element {
+        width: 182px;
+        height: 36px;
+      }
+    }
+
+    .skeleton-element.chart {
+      height: 140px;
+      margin-bottom: 12px;
+    }
+  }
+
+  .statistics-info {
+    .top-k-info-header {
+      display: flex;
+      align-items: center;
+      justify-content: space-between;
+      margin-bottom: 12px;
+
+      .label-item {
+        font-size: 12px;
+        line-height: 20px;
+        color: #4d4f56;
+      }
+    }
+
+    .integer-statics-info {
+      display: flex;
+      flex-wrap: wrap;
+      align-content: space-between;
+      justify-content: space-between;
+      height: 76px;
+      margin-bottom: 15px;
+
+      .integer-item {
+        width: 182px;
+        height: 36px;
+        line-height: 36px;
+        text-align: center;
+        background: #f5f7fa;
+        border-radius: 2px;
+
+        .label {
+          font-size: 12px;
+          color: #4d4f56;
+        }
+
+        .value {
+          margin-left: 8px;
+          font-size: 12px;
+          font-weight: 700;
+          color: #4d4f56;
+        }
+      }
+    }
+
+    .top-k-chart-title {
+      display: flex;
+      justify-content: space-between;
+      gap: 20px;
+      margin-bottom: 4px;
+      line-height: 22px;
+
+      .title {
+        flex-shrink: 0;
+        font-size: 14px;
+        color: #313238;
+      }
+
+      .time-range {
+        overflow: hidden;
+        text-overflow: ellipsis;
+        white-space: nowrap;
+        font-size: 12px;
+        color: #4d4f56;
+      }
+    }
+  }
+}
