+++ bkmonitor/webpack/src/trace/pages/trace-explore/components/statistics-list/topk-field-list.scss
@@ -1,0 +1,81 @@
+.trace-explore-dimension-statistics-popover {
+  .skeleton-wrap {
+    .skeleton-element {
+      height: 28px;
+      margin-top: 12px;
+    }
+  }
+
+  .empty-text {
+    font-size: 12px;
+  }
+}
+
+.top-k-list {
+  .top-k-list-item {
+    display: flex;
+    margin-top: 12px;
+
+    .filter-tools {
+      position: relative;
+      margin: 2px 6px 0 0;
+      color: #3a84ff;
+
+      .icon-monitor {
+        cursor: pointer;
+      }
+
+      .icon-sousuo- {
+        margin-left: 8px;
+      }
+    }
+
+    .progress-content {
+      flex: 1;
+      overflow: hidden;
+
+      .info-text {
+        display: flex;
+        align-items: center;
+        justify-content: space-between;
+        height: 20px;
+        margin-bottom: 2px;
+
+        .field-name {
+          flex: 1;
+          overflow: hidden;
+          text-overflow: ellipsis;
+          font-size: 12px;
+          color: #313238;
+          white-space: nowrap;
+
+          .sub-name {
+            color: #757880;
+          }
+        }
+
+        .counts {
+          font-size: 12px;
+          color: #5ab8a8;
+
+          .progress-count {
+            margin-left: 4px;
+            font-weight: 700;
+          }
+        }
+      }
+    }
+  }
+}
+
+.slide-dimension-filter-overflow-tips {
+  font-size: 12px;
+}
+
+.slide-dimension-filter-overflow-tips,
+.statistics-top-k-item-tooltips-wrap-popover {
+  max-width: 400px;
+  word-break: break-word;
+  overflow-wrap: break-word;
+  white-space: normal;
+}
